new cmake infrastructure and start of fixing warnings in llarm-emu

+++ llarm-emu/src/cpu/core/cycle/fetch.cpp
@@ -17,7 +17,7 @@ arm_fetch_struct FETCH::arm_fetch() {
     mem_read_struct access = memory.read(reg.read(id::reg::PC), 4, id::access_type::INSTRUCTION_FETCH);
     
     if (globals.is_little_endian) {
-        access.value = util::swap_endianness<u32>(access.value);
+        access.value = util::swap_endianness<u32>(u32(access.value));
     }
 
     if (access.has_failed) {
@@ -39,7 +39,7 @@ thumb_fetch_struct FETCH::thumb_fetch() {
     mem_read_struct access = memory.read(reg.read(id::reg::PC), 2, id::access_type::INSTRUCTION_FETCH);
 
     if (globals.is_little_endian) {
-        access.value = util::swap_endianness<u16>(access.value);
+        access.value = util::swap_endianness<u16>(u16(access.value));
     }
 
     if (access.has_failed) {
